SDL: Move class definition out of header.
